+++ source/renderer/app/themes/daedalus/cardano.js
@@ -749,6 +749,7 @@ export const CARDANO_THEME_OUTPUT = {
       'rgba(32, 34, 37, .1)',
     '--theme-staking-stake-pool-tooltip-neutral-text-color': '#5e6066',
     '--theme-staking-stake-pool-tooltip-experimental-icon-color': '#5e6066',
+    '--theme-staking-stake-pool-tooltip-id-background-color': '#fafbfc',
   },
   staking: {
     '--theme-staking-content-background-color': '#ffffff',